+++ tokobuku/package.json
@@ -11,13 +11,15 @@
   },
   "dependencies": {
     "@react-native-community/masked-view": "^0.1.10",
+    "@react-navigation/bottom-tabs": "^5.5.2",
     "@react-navigation/native": "^5.5.1",
     "@react-navigation/stack": "^5.5.1",
+    "native-base": "^2.13.12",
     "react": "16.11.0",
     "react-native": "^0.62.2",
     "react-native-gesture-handler": "^1.6.1",
     "react-native-reanimated": "^1.9.0",
-    "react-native-safe-area-context": "^3.0.4",
+    "react-native-safe-area-context": "^3.0.5",
     "react-native-screens": "^2.8.0",
     "react-native-svg": "^12.1.0",
     "react-native-svg-transformer": "^0.14.3"
