+++ tokobuku/src/App.js
@@ -1,14 +1,15 @@
 import { NavigationContainer } from '@react-navigation/native';
 import React from 'react';
 import { StyleSheet } from 'react-native';
-import Router from './router';
+import Router from './router/RegisterLoginRouter';
+import InitialRouter from './router/InitialRouter';
 
 // KERJAKAN BAGIAN APP.JS
 
 const App = () => {
   return (
     <NavigationContainer>
-		<Router />
+      <InitialRouter/>
     </NavigationContainer>
   )
 }
